+++ changelog.md
@@ -40,3 +40,4 @@
 - removed `teaser` from `article-page`
 - removed `author` from `article-page`
 - removed `author`
+- removed `headline` from `article-overview-page`